Moved bbosh, data, and tcp modules into their own repos netx.bbosh, netx.data, and netx.tcp respectively. Moved the URLFactory and URLStreamHandlerFactorySpi to org.kaazing.netx package in the netx repo. Deleted the api module.
