+++ README.md
@@ -1,4 +1,4 @@
-Kaazing net.api
-========
+Kaazing netx
+============
 
 Utility java.net functions to make life simpler
