+++ src/main/java/org/kaazing/netx/URLStreamHandlerFactorySpi.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.kaazing.net;
+package org.kaazing.netx;
 
 import java.net.URLStreamHandlerFactory;
 import java.util.Collection;
